+++ cms/src/hooks/useRefreshToken.tsx
@@ -1,13 +1,11 @@
-import { AxiosResponse } from "axios";
 import axiosDefault from "../config/axios.config";
 import { AuthSuccessResponse } from "../interfaces/responses";
 
 const useRefreshToken = () => {
   const refresh = async () => {
     console.log("refresh running...");
-    const response = await axiosDefault.post<AuthSuccessResponse>(
-      "/auth/refresh"
-    );
+    const response =
+      await axiosDefault.post<AuthSuccessResponse>("/auth/refresh");
     // console.log(response.)
     return response.data;
   };