+++ cms/src/routes/_protected/dashboard.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import Dashboard from "../../../pages/dashboard/Dashboard";
+import Dashboard from "../../pages/dashboard/Dashboard";
 
 export const Route = createFileRoute("/_protected/dashboard")({
   component: Dashboard,